docs: Clarify the documentation now that there are more targets
Since the addition of other targets, it is necessary to be more explicit
about the way yotta can be used to build BLE examples.

Furthermore, some of the readme files referred to hex file names
specific to the Nordic targets, and this change fixes that.

Finally, a slightly clarified 'success criteria' has been added to each
example to assist people following the instructions in ascertaining that
they have done so correctly.

Signed-off-by: Jonathan Austin <jonathan.austin@arm.com>